refactor(tests): move utility functions to separate file, use new vm every time
